pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.backends.legacy:BuildBackend"

[project]
name = "data-cleaner-rl"
version = "1.0.0"
description = "OpenEnv-compatible RL environment for broken pandas pipelines"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "gradio>=4.0.0",
    "torch>=2.0.0",
    "matplotlib>=3.8.0",
    "openenv-core @ git+https://github.com/meta-pytorch/OpenEnv.git",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "httpx>=0.25.0"]

[project.scripts]
server = "src.server:main"
